backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
supabase==2.5.0
google-generativeai==0.7.2
pydantic==2.7.4
slowapi==0.1.9
cachetools==5.3.3
python-jose[cryptography]==3.3.0
httpx==0.27.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
spacy>=3.7.0
sentence-transformers>=2.6.0
openai>=1.30.0
xgboost>=2.0.0
